Hands_on_Lab/Day_2/Task_1/requirements.txt
# PDF RAG — Python Dependencies
# Install with: pip install -r requirements.txt

# Web framework
flask
flask-cors

# PDF processing
pypdf

# Text chunking
langchain
langchain-text-splitters

# Vector database (local, persistent)
chromadb

# Embeddings (free, runs locally via sentence-transformers)
sentence-transformers

# LLM inference (free Groq API)
groq

# Environment variable management
python-dotenv
